+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 -   Log process output to a separate channel than extension output
+-   Change status bar text strings to be capitalized similar to other extensions
+-   Cursor follow now says "Following Cursor" instead of "Stop Following Cursor"
 
 ## 0.14.0 - 2024-04-20